requirements.txt
chromadb==0.4.24
qdrant-client==1.9.1
sentence-transformers==3.0.1
transformers==4.41.2
torch>=2.1
ragas==0.1.10
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.4
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain==0.2.10
ollama==0.2.0   # (optional if you prefer the client)
